chore: clean up UiState and UiEvent
