--- shared/feature/details/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/details/genre/GenreUiGenre.kt
@@ -1,8 +1,0 @@
-package ly.david.musicsearch.shared.feature.details.genre
-
-import com.slack.circuit.runtime.CircuitUiEvent
-
-internal sealed interface GenreUiGenre : CircuitUiEvent {
-    data object NavigateUp : GenreUiGenre
-    data object ForceRefresh : GenreUiGenre
-}
