+++ shared/feature/details/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/details/label/LabelPresenter.kt
@@ -2,6 +2,7 @@ package ly.david.musicsearch.shared.feature.details.label
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.Stable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -9,19 +10,24 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import com.slack.circuit.foundation.NavEvent
 import com.slack.circuit.foundation.onNavEvent
+import com.slack.circuit.runtime.CircuitUiEvent
+import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.presenter.Presenter
 import ly.david.musicsearch.core.logging.Logger
 import ly.david.musicsearch.core.models.getNameWithDisambiguation
 import ly.david.musicsearch.core.models.history.LookupHistory
 import ly.david.musicsearch.core.models.label.LabelScaffoldModel
+import ly.david.musicsearch.core.models.network.MusicBrainzEntity
 import ly.david.musicsearch.data.common.network.RecoverableNetworkException
 import ly.david.musicsearch.domain.history.usecase.IncrementLookupHistory
 import ly.david.musicsearch.domain.label.LabelRepository
 import ly.david.ui.common.relation.RelationsPresenter
 import ly.david.ui.common.relation.RelationsUiEvent
+import ly.david.ui.common.relation.RelationsUiState
 import ly.david.ui.common.release.ReleasesByEntityPresenter
 import ly.david.ui.common.release.ReleasesByEntityUiEvent
+import ly.david.ui.common.release.ReleasesByEntityUiState
 import ly.david.ui.common.screen.DetailsScreen
 
 internal class LabelPresenter(
@@ -155,4 +161,29 @@ internal class LabelPresenter(
             eventSink = ::eventSink,
         )
     }
+}
+
+@Stable
+internal data class LabelUiState(
+    val title: String,
+    val isError: Boolean,
+    val label: LabelScaffoldModel?,
+    val tabs: List<LabelTab>,
+    val selectedTab: LabelTab,
+    val query: String,
+    val releasesByEntityUiState: ReleasesByEntityUiState,
+    val relationsUiState: RelationsUiState,
+    val eventSink: (LabelUiEvent) -> Unit,
+) : CircuitUiState
+
+internal sealed interface LabelUiEvent : CircuitUiEvent {
+    data object NavigateUp : LabelUiEvent
+    data object ForceRefresh : LabelUiEvent
+    data class UpdateTab(val tab: LabelTab) : LabelUiEvent
+    data class UpdateQuery(val query: String) : LabelUiEvent
+    data class ClickItem(
+        val entity: MusicBrainzEntity,
+        val id: String,
+        val title: String?,
+    ) : LabelUiEvent
 }
