+++ shared/feature/details/src/commonMain/kotlin/ly/david/musicsearch/shared/feature/details/recording/RecordingPresenter.kt
@@ -2,6 +2,7 @@ package ly.david.musicsearch.shared.feature.details.recording
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.Stable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -9,20 +10,25 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import com.slack.circuit.foundation.NavEvent
 import com.slack.circuit.foundation.onNavEvent
+import com.slack.circuit.runtime.CircuitUiEvent
+import com.slack.circuit.runtime.CircuitUiState
 import com.slack.circuit.runtime.Navigator
 import com.slack.circuit.runtime.presenter.Presenter
 import ly.david.musicsearch.core.logging.Logger
 import ly.david.musicsearch.core.models.artist.getDisplayNames
 import ly.david.musicsearch.core.models.getNameWithDisambiguation
 import ly.david.musicsearch.core.models.history.LookupHistory
+import ly.david.musicsearch.core.models.network.MusicBrainzEntity
 import ly.david.musicsearch.core.models.recording.RecordingScaffoldModel
 import ly.david.musicsearch.data.common.network.RecoverableNetworkException
 import ly.david.musicsearch.domain.history.usecase.IncrementLookupHistory
 import ly.david.musicsearch.domain.recording.RecordingRepository
 import ly.david.ui.common.relation.RelationsPresenter
 import ly.david.ui.common.relation.RelationsUiEvent
+import ly.david.ui.common.relation.RelationsUiState
 import ly.david.ui.common.release.ReleasesByEntityPresenter
 import ly.david.ui.common.release.ReleasesByEntityUiEvent
+import ly.david.ui.common.release.ReleasesByEntityUiState
 import ly.david.ui.common.screen.DetailsScreen
 
 internal class RecordingPresenter(
@@ -159,4 +165,30 @@ internal class RecordingPresenter(
             eventSink = ::eventSink,
         )
     }
+}
+
+@Stable
+internal data class RecordingUiState(
+    val title: String,
+    val subtitle: String,
+    val isError: Boolean,
+    val recording: RecordingScaffoldModel?,
+    val tabs: List<RecordingTab>,
+    val selectedTab: RecordingTab,
+    val query: String,
+    val releasesByEntityUiState: ReleasesByEntityUiState,
+    val relationsUiState: RelationsUiState,
+    val eventSink: (RecordingUiEvent) -> Unit,
+) : CircuitUiState
+
+internal sealed interface RecordingUiEvent : CircuitUiEvent {
+    data object NavigateUp : RecordingUiEvent
+    data object ForceRefresh : RecordingUiEvent
+    data class UpdateTab(val tab: RecordingTab) : RecordingUiEvent
+    data class UpdateQuery(val query: String) : RecordingUiEvent
+    data class ClickItem(
+        val entity: MusicBrainzEntity,
+        val id: String,
+        val title: String?,
+    ) : RecordingUiEvent
 }
